requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.9.0
blinker==1.4
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4 ; python_version > '3.0'
click==7.1.2
dnspython==1.16.0
email-validator==1.0.5
flask-mail==0.9.1
flask-wtf==0.14.3
flask==1.1.2
google-api-core[grpc]==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
numpy==1.18.3
pandas==1.0.3
pdfminer-six==20200402
protobuf==3.11.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.9.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
urllib3==1.25.9
werkzeug==1.0.1
wtforms==2.3.1
xlrd==1.2.0
